generalize more...ish
